refactor: 사용하지 않는 콘솔 연관 클래스 제거
